no decimals

+++ UniversalRecallTracker/Program.cs
@@ -190,7 +190,7 @@ namespace UniversalRecallTracker
                         _sprite.X - rect.Width - GapTextBar,
                         _sprite.Y - rect.Height / 2 + (_sprite.Height * Program.Instance().BarScale) / 2);
                 },
-                TextUpdate = () => "(" + hero.HealthPercent + "%)"
+                TextUpdate = () => "(" + (int) hero.HealthPercent + "%)"
             };
             _healthText.Add(1);
             Render.Text heroText = new Render.Text(0, 0, hero.ChampionName, TextFont.Description.Height, Color.White)
@@ -232,7 +232,7 @@ namespace UniversalRecallTracker
                     case Packet.S2C.Teleport.Status.Finish:
                         int colorIndex = (int) ((_hero.HealthPercent / 100) * 255);
                         string color = (255 - colorIndex).ToString("X2") + colorIndex.ToString("X2") + "00";
-                        Program.Instance().Notify(_hero.ChampionName + " has recalled with <font color='#" + color + "'>" + _hero.HealthPercent + "&#37; HP</font>");
+                        Program.Instance().Notify(_hero.ChampionName + " has recalled with <font color='#" + color + "'>" + (int) _hero.HealthPercent + "&#37; HP</font>");
                         _active = false;
                         break;
                     case Packet.S2C.Teleport.Status.Abort:
